add superclass for pantsus
